Config dev environment

+++ nestjs-task-management/src/tasks/tasks.service.ts
@@ -18,8 +18,8 @@ export class TasksService {
     private taskRepository: TaskRepository
   ) { }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
     if (!found) {
       throw new NotFoundException(`Task with id ${id} not found!`);
@@ -32,24 +32,22 @@ export class TasksService {
     return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with id ${id} not found!`);
     }
   }
 
-  async updateTask(id: number, updateTaskDto: UpdateTaskDto): Promise<Task> {
-    let task = await this.getTaskById(id);
+  async updateTask(id: number, updateTaskDto: UpdateTaskDto, user: User): Promise<Task> {
+    let task = await this.getTaskById(id, user);
     task.status = updateTaskDto.status;
     await task.save();
     return task;
   }
 
   async getTasks(filterDto: FilterTaskDto, user: User): Promise<Task[]> {
-    console.log("service", user)
-
     return this.taskRepository.getTasks(filterDto, user);
   }
